Updated Timing Endpoint
Updated the firmware to fix known errors related to the timing endpoint module, as well as introduced some debug probes related to this module. The Design DOES NOT meet timing because of one critical path that is related to one clock that only feeds one debug probe, therefore we would like to assume that it is safe to ignore it this time.