Updated files to address design rebuilding
Changed all the scripts that build the design to address the automatic creation of different modules and files. Previously, the scripts would scan to see if the files already existed, if so, it would use those files when generating the new overlays, therefore, older IPs or xci files would be used with new modules and conflict would emerge. To address this issue, the strategy was to manually delete the files, but this defies the idea is to keep the design as automatic as possible. Now, every time the batch file is run, a clean version of the repo is generated first before generating all the design files.
Updated the README to fit the name of the design used in this branch
Added -notrace argument to the source commands to clean the output in TCL console